+++ src/WOTRMultiplayer.Networking/Messages/Contracts/NetworkLevelingPortrait.cs
@@ -1,0 +1,17 @@
+﻿using ProtoBuf;
+
+namespace WOTRMultiplayer.Networking.Messages.Contracts
+{
+    [ProtoContract]
+    public class NetworkLevelingPortrait
+    {
+        [ProtoMember(1)]
+        public string Name { get; set; }
+
+        [ProtoMember(2)]
+        public string CustomId { get; set; }
+
+        [ProtoMember(3)]
+        public string Category { get; set; }
+    }
+}