+++ src/WOTRMultiplayer.Networking/Messages/MessageTypes.cs
@@ -124,6 +124,7 @@
             NotifyDialogPopupClosed,
             NotifyInventoryItemUsed,
             NotifyLevelingMythicClassSelected,
+            NotifyLevelingPortraitSelected,
         }
 
         public enum Request