+++ src/WOTRMultiplayer/MP/Entities/Leveling/NetworkLevelingPortrait.cs
@@ -1,0 +1,11 @@
+﻿namespace WOTRMultiplayer.MP.Entities.Leveling
+{
+    public class NetworkLevelingPortrait
+    {
+        public string Name { get; set; }
+
+        public string CustomId { get; set; }
+
+        public string Category { get; set; }
+    }
+}